@@ -9,21 +9,19 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^4.2.5",
-    "@angular/cli": "^1.2.0",
-    "@angular/common": "^4.2.5",
-    "@angular/compiler": "^4.2.5",
-    "@angular/compiler-cli": "^4.2.5",
-    "@angular/core": "^4.2.5",
-    "@angular/forms": "^4.2.5",
-    "@angular/http": "^4.2.5",
-    "@angular/platform-browser": "^4.2.5",
-    "@angular/platform-browser-dynamic": "^4.2.5",
-    "@angular/platform-server": "^4.2.5",
-    "@angular/router": "^4.2.5",
-    "angularfire2": "^4.0.0-rc0",
+    "@angular/animations": "^4.3.x",
+    "@angular/cli": "^1.3.0",
+    "@angular/common": "^4.3.x",
+    "@angular/compiler": "^4.3.x",
+    "@angular/compiler-cli": "^4.3.x",
+    "@angular/core": "^4.3.x",
+    "@angular/forms": "^4.3.x",
+    "@angular/http": "^4.3.x",
+    "@angular/platform-browser": "^4.3.x",
+    "@angular/platform-browser-dynamic": "^4.3.x",
+    "@angular/platform-server": "^4.3.x",
+    "@angular/router": "^4.3.x",
     "core-js": "^2.4.1",
-    "firebase": "^4.1.3",
     "rxjs": "^5.4.1",
     "ts-helpers": "^1.1.1",
     "typescript": "^2.4.1",
