@@ -4,13 +4,5 @@
 // The list of which env maps to which file can be found in `angular-cli.json`.
 
 export const environment = {
-	production: false,
-	firebase: {
-		apiKey: "AIzaSyA2Zipd8-hLczQw6xTK9nqvZp9yoYZqE5I",
-		authDomain: "foosball-11e4e.firebaseapp.com",
-		databaseURL: "https://foosball-11e4e.firebaseio.com",
-		projectId: "foosball-11e4e",
-		storageBucket: "",
-		messagingSenderId: "913793052747"
-	}
+	production: false
 };
